requirements.txt
# MQTT to Telegram Rankings Bot with Execution Layer Dependencies

# MQTT client library
paho-mqtt>=1.6.0,<2.0.0

# Telegram bot library
python-telegram-bot>=20.0,<21.0

# Environment configuration
python-dotenv>=1.0.0

# Hummingbot API client
# Note: This needs to be installed from the Hummingbot repository
# pip install git+https://github.com/hummingbot/hummingbot-api-client.git
# Or specify local path if you have it locally
#hummingbot-api-client>=0.1.0

# Data processing and analysis
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# YAML configuration handling
PyYAML>=6.0,<7.0

# HTTP requests for API calls
requests>=2.31.0,<3.0.0
aiohttp>=3.8.0,<4.0.0

# Async support
asyncio-mqtt>=0.13.0,<1.0.0

# Date and time utilities
python-dateutil>=2.8.2,<3.0.0

# Optional: For enhanced logging
colorlog>=6.7.0,<7.0.0

# Optional: For data validation
pydantic>=2.0.0,<3.0.0